[coordinates/readers] Added reader for HDF5 format
Because a HDF5 file can contain multiple distinct data sets,
we provide a selection functionality with regular expression
support. All the matching data sets (their absolute path) will be
considered as unique trajectories.

At the moment the selection can only be a single string, but we
should add a mapping for {filenames: selection} in the future
in order to be more flexible.