[base/Estimator] added decorator/ctx manager for const input data checking
For now we only allow input for fit to be of type ndarray or list/tuple of ndarray
in order to check constness during fit/partial_fit calls.

Newly introduced a Estimator (or derived) class flag *_MUTABLE_INPUT_DATA* to control
this checking (e.g. can be switched off, if really undesired).