Update parallel function according to documentation (#272)
* change the function of n_jobs to match documentation

In scikit-learn, we can use n_jobs to specify how many cores we are using, generally -1 for all. This is also in our documentation (for example in clustering methods). However, it doesn't match here and will raise error when using -1.

* Update AUTHORS

Add myself to `AUTHORS` as contributors

* Reconstruct logic of handle_n_jobs

First determine if using None, then determine if using -1, at last check if it's a positive integer.
Add unit test for handle_n_jobs function